Renamed Jsoncpp to prefab-depdency, updated README.md
