--- prefab/json-parse/settings.gradle
@@ -1,2 +1,0 @@
-rootProject.name='json-parse'
-include ':app'
