+++ prefab/prefab-dependecy/app/src/main/cpp/Android.mk
