+++ prefab/prefab-dependecy/app/src/main/cpp/Application.mk
