+++ prefab/prefab-dependecy/app/src/main/cpp/java_interop.h
@@ -54,4 +54,4 @@ struct Convert<jstring> {
   }
 };
 
-}  // namespace jsonparse::jni
+}  // namespace com.example.prefabdependency::jni
