+++ prefab/prefab-dependecy/app/src/main/cpp/logging.cpp
@@ -14,13 +14,27 @@
  * limitations under the License.
  */
 
-#pragma once
+#include "logging.h"
 
 #include <jni.h>
+#include <stdarg.h>
+#include <stdio.h>
+#include <stdlib.h>
 
 namespace jsonparse::logging {
 
 [[noreturn, gnu::format(printf, 2, 3)]] void FatalError(JNIEnv* env,
-                                                        const char* fmt, ...);
+                                                        const char* fmt, ...) {
+  va_list ap;
+  va_start(ap, fmt);
+  char* msg = nullptr;
+  vasprintf(&msg, fmt, ap);
+  va_end(ap);
 
-}  // namespace jsonparse::logging
+  env->FatalError(msg);
+  // env->FatalError() is specified to not return, but the function is not
+  // annotated with the noreturn attribute. abort() just in case.
+  abort();
+}
+
+}  // namespace com.example.prefabdependency::logging
